Удалить head в односвязном списке

+++ src/test/java/ru/job4j/collection/ForwardLinkedTest.java
@@ -27,7 +27,7 @@ public class ForwardLinkedTest {
         ForwardLinked<Integer> linked = new ForwardLinked<>();
         linked.add(1);
         linked.add(2);
-        assertThat(linked.deleteFirst(), is(1));
+        assertThat(linked.deleteFirst().intValue(), is(1));
         Iterator<Integer> it = linked.iterator();
         assertThat(it.next(), is(2));
     }
